refactor: 豆のクエリ用MapperをCoffeeBeanQueryMapperとしてadmin-apiに分離
commonの集約用CoffeeBeanMapperに混在していたクエリメソッド
(findListRows/findDetailRowsById/countAll)とRow entityを、
cs-apiと同様にadmin-apiモジュール内の専用CoffeeBeanQueryMapperへ移動。
commonのCoffeeBeanMapperはコマンド側専用に戻す。
