+++ backend/admin-api/src/main/kotlin/com/mametosho/admin/AdminApiApplication.kt
@@ -5,7 +5,7 @@ import org.springframework.boot.autoconfigure.SpringBootApplication
 import org.springframework.boot.runApplication
 
 @SpringBootApplication(scanBasePackages = ["com.mametosho"])
-@MapperScan("com.mametosho.infrastructure.persistence.mybatis.mapper")
+@MapperScan("com.mametosho.infrastructure.persistence.mybatis.mapper", "com.mametosho.admin.infrastructure.persistence.mybatis")
 class AdminApiApplication
 
 fun main(args: Array<String>) {
