+++ backend/admin-api/src/main/kotlin/com/mametosho/admin/infrastructure/persistence/query/CoffeeBeanQueryServiceImpl.kt
@@ -4,18 +4,18 @@ import com.mametosho.admin.application.query.CoffeeBeanQueryService
 import com.mametosho.admin.application.query.result.CoffeeBeanDetailResult
 import com.mametosho.admin.application.query.result.CoffeeBeanSummaryResult
 import com.mametosho.admin.application.result.PagedResult
-import com.mametosho.infrastructure.persistence.mybatis.mapper.CoffeeBeanMapper
+import com.mametosho.admin.infrastructure.persistence.mybatis.mapper.CoffeeBeanQueryMapper
 import org.springframework.stereotype.Service
 
 @Service
 class CoffeeBeanQueryServiceImpl(
-    private val coffeeBeanMapper: CoffeeBeanMapper,
+    private val coffeeBeanQueryMapper: CoffeeBeanQueryMapper,
 ) : CoffeeBeanQueryService {
 
     override fun findList(page: Int, size: Int): PagedResult<CoffeeBeanSummaryResult> {
         val offset = page * size
-        val rows = coffeeBeanMapper.findListRows(size, offset)
-        val totalCount = coffeeBeanMapper.countAll()
+        val rows = coffeeBeanQueryMapper.findListRows(size, offset)
+        val totalCount = coffeeBeanQueryMapper.countAll()
 
         val items = rows.map { row ->
             CoffeeBeanSummaryResult(
@@ -43,7 +43,7 @@ class CoffeeBeanQueryServiceImpl(
     }
 
     override fun findDetail(id: String): CoffeeBeanDetailResult? {
-        val rows = coffeeBeanMapper.findDetailRowsById(id)
+        val rows = coffeeBeanQueryMapper.findDetailRowsById(id)
         if (rows.isEmpty()) return null
 
         val first = rows.first()
